+++ src/Core/WynajemMaszyn.Application/Features/Rollers/Command/CreateRollers/CreateRollerCommandHandler.cs
@@ -33,7 +33,7 @@ namespace WynajemMaszyn.Application.Features.Rollers.Command.CreateRollers
 
             var roller = new Roller
             {
-                UserId = (int)userId,
+                UserId = userId,
                 Name = request.Name,
                 ProductionYear = request.ProductionYear,
                 OperatingHours = request.OperatingHours,
